refactor: update Node.js version in Dockerfiles and improve environment configuration
