+++ Dockerfile.prod
@@ -1,7 +1,7 @@
 # ----------------------------
 # Copy & Install dependencies
 # ----------------------------
-FROM node:21-alpine3.19 as deps
+FROM node:22-alpine3.22 AS deps
 WORKDIR /usr/src/app
 COPY package*.json ./
 RUN npm install
@@ -9,7 +9,7 @@ RUN npm install
 # -----------------
 # Copy & Build App
 # -----------------
-FROM node:21-alpine3.19 as build
+FROM node:22-alpine3.22 AS build
 WORKDIR /usr/src/app
 COPY --from=deps /usr/src/app/node_modules/ ./node_modules
 COPY . .
@@ -19,7 +19,7 @@ RUN npm ci -f --only=production && npm cache clean --force
 # --------------------
 # Create image Docker
 # --------------------
-FROM node:21-alpine3.19 as prod
+FROM node:22-alpine3.22 AS prod
 WORKDIR /usr/src/app
 COPY --from=build /usr/src/app/node_modules/ ./node_modules
 COPY --from=build /usr/src/app/dist ./dist
