+++ src/orders/orders.controller.ts
@@ -9,7 +9,7 @@ import {
   Post,
   Query,
 } from '@nestjs/common';
-import enviroment from '../config/configuration';
+import enviroment from '../config/env.config';
 import { ClientProxy, RpcException } from '@nestjs/microservices';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { PaginationDto } from 'src/common';
