+++ src/products/products.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ProductsController } from './products.controller';
 import { ClientsModule, Transport } from '@nestjs/microservices';
-import enviroment from '../config/configuration';
+import enviroment from '../config/env.config';
 
 @Module({
   controllers: [ProductsController],
